Move buffer_pool to zenoh_utils
