+++ rmw_zenoh_cpp/CMakeLists.txt
@@ -49,7 +49,6 @@ add_library(rmw_zenoh_cpp SHARED
   src/detail/type_support_common.cpp
   src/detail/zenoh_config.cpp
   src/detail/zenoh_utils.cpp
-  src/detail/buffer_pool.cpp
   src/rmw_event.cpp
   src/rmw_get_network_flow_endpoints.cpp
   src/rmw_get_node_info_and_types.cpp
