+++ rmw_zenoh_cpp/src/detail/zenoh_utils.hpp
@@ -20,13 +20,18 @@
 #include <array>
 #include <chrono>
 #include <cstddef>
+#include <cstdlib>
 #include <functional>
+#include <mutex>
 #include <optional>
 #include <utility>
 #include <variant>
 #include <vector>
 
+#include "rcutils/allocator.h"
+#include "rcutils/env.h"
 #include "rmw/types.h"
+#include "logging_macros.hpp"
 
 namespace rmw_zenoh_cpp
 {
@@ -103,6 +108,36 @@ struct ShmContext
   size_t msgsize_threshold;
 
   ShmContext(size_t alloc_size, size_t msgsize_threshold);
+};
+
+///=============================================================================
+class BufferPool
+{
+public:
+  struct Buffer
+  {
+    uint8_t * data;
+    size_t size;
+  };
+
+  BufferPool();
+
+  ~BufferPool();
+
+  Buffer allocate(size_t size);
+
+  void deallocate(Buffer buffer);
+
+private:
+  std::vector<Buffer> buffers_;
+  std::mutex mutex_;
+  size_t max_size_;
+  size_t size_;
+  // NOTE(fuzzypixelz): Pooled buffers are recycled with the expectation that they would reside in
+  // cache, thus this this value should be comparable to the size of a modern CPU cache. The default
+  // value (16 MiB) is relatively conservative as CPU cache sizes range from a few MiB to a few
+  // hundred MiB.
+  const size_t DEFAULT_MAX_SIZE = 16 * 1024 * 1024;
 };
 }  // namespace rmw_zenoh_cpp
